requirements.txt
pyrogram
python-dotenv
beautifulsoup4
requests
reportlab
pymongo
certifi